Add files via upload
Analysis has been done of the Condition of disease "Schizophrenia" in INDIA over the past few years in different states and among different age groups and what policies can be implemented to overcome this issue. Questions like, what needs to be done on the part of the government, on the part of citizens of the country, and more importantly as a parent, have been answered.